JB: WIP on user service logic, but schema may change

+++ src/main/java/com/homebab/server/repository/ItemRepository.java
@@ -1,8 +1,8 @@
 package com.homebab.server.repository;
 
-import com.homebab.server.model.Item;
+import com.homebab.server.domain.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemRepository extends JpaRepository<Item, Integer> {
+public interface ItemRepository extends JpaRepository<Item, Long> {
 
 }
