+++ src/main/java/com/homebab/server/repository/RecipeRepository.java
@@ -1,13 +1,9 @@
 package com.homebab.server.repository;
 
-import com.homebab.server.model.Recipe;
+import com.homebab.server.domain.Recipe;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
-import java.util.Collection;
-
-public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
+public interface RecipeRepository extends JpaRepository<Recipe, Long> {
 
 //    @Query(value = "SELECT r.*"
 //        + " FROM recipes r"
