+++ src/main/java/com/homebab/server/repository/UserRepository.java
@@ -1,8 +1,8 @@
 package com.homebab.server.repository;
 
-import com.homebab.server.model.User;
+import com.homebab.server.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer>  {
+public interface UserRepository extends JpaRepository<User, Long>  {
 
 }
